Compile some frame buffer drivers into the kernel instead of building them as modules
On OldWorld Macs, frame buffer drivers cannot get the display resolution, often defaulting to a resolution that doesn't exist especially on laptops. To fix this, the resolution can be specified on
the kernel command line for each driver (e.g. video=atyfb:vmode:10,cmode:32 for 800x600@60-32), but in the case of atyfb at least, this functionality is broken when it is compiled as a module.